Switch off trying to find model instances
Been spending over 80h on trying to migrate my database to the heroku postgres database. All tables are migrated but they are all empty. It's not working and at this point I can't waste anymore time trying to solve it, I'm gonna write everything in the database manually in the postgres database and need to remove some searches for model instances to be able to access the admin page at all.